refactor: Move ToC2html functionality to sidebar
